Nytt endepunkt for å hente aktive varsler
